PasswordRecoveryService.java
package Trial;

import java.util.Date;
import java.util.Optional;

public class PasswordRecoveryService {

    // Method to start the password recovery process
    public void initiatePasswordRecovery(String emailOrMobile) {
        // Verify the email or mobile number
        // Generate and send a reset token via email or SMS
    }

    // Method to check if the token is valid
    public boolean validateResetToken(String token) {
        // Check if the token is valid and not expired
        return false;  // This is a placeholder
    }

    // Method to reset the password
    public void resetPassword(String token, String newPassword) {
        // Validate the token
        // Ensure the new password is secure
        // Update the user's password in the database
    }

    // Helper methods could be added here, such as generating tokens and sending them
}
